Modernize for Python 3.12, pandas 2.x, numpy 2.x, seaborn 0.13
Source fixes:
- Replace removed pandas APIs: Series.iteritems -> items,
  DataFrame.append -> pd.concat, nested-renamer .agg(dict) ->
  named aggregation, pd.Float64Index -> pd.Index(dtype=float64),
  groupby(axis=1) -> transpose groupby, applymap -> vectorized
  comparison, resample('M') -> resample('ME')
- Replace removed seaborn distplot with histplot; rewrite
  plot_return_quantiles box/swarm plots in long form (seaborn 0.13
  cannot align DatetimeIndex with aggregate_returns' MultiIndex)
- Cast perf stats table to object dtype before writing percent
  strings (pandas 3.0 forbids incompatible-dtype setitem)
- Fix set_yticks/set_yticklabels count mismatch in
  plot_round_trip_lifetimes when fewer symbols than disp_amount
- Set MultiIndex level names via set_names in capacity.py
- Select vwap columns before groupby.apply; keep turnover float

Packaging and tests:
- Depend on empyrical-reloaded; require Python >= 3.9 with modern
  minimum versions
- Migrate tests from nose/nose-parameterized/runipy to
  pytest/parameterized/nbclient; pandas.util.testing -> pandas.testing
- Gate legacy example-notebook execution behind PYFOLIO_TEST_NBS=1
- Update stale test expectations (read_csv squeeze kwarg,
  assertRaisesRegexp, string version comparison, 0/0 rolling sharpe)

+++ pyfolio/txn.py
@@ -202,5 +202,5 @@ def get_turnover(positions, transactions, denominator='AGB'):
 
     denom.index = denom.index.normalize()
     turnover = traded_value.div(denom, axis='index')
-    turnover = turnover.fillna(0)
+    turnover = turnover.astype(float).fillna(0)
     return turnover
